+++ qupath-gui-fx/src/main/java/qupath/lib/gui/panes/PreferencePane.java
@@ -24,7 +24,12 @@
 package qupath.lib.gui.panes;
 
 import java.io.File;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
+import java.util.function.Function;
+
 import org.controlsfx.control.PropertySheet;
 import org.controlsfx.control.PropertySheet.Item;
 import org.controlsfx.control.PropertySheet.Mode;
@@ -41,18 +46,22 @@ import javafx.beans.property.Property;
 import javafx.beans.value.ObservableValue;
 import javafx.collections.ObservableList;
 import javafx.scene.Node;
+import javafx.scene.control.ComboBox;
 import javafx.scene.control.TextField;
 import javafx.scene.control.Tooltip;
 import javafx.scene.paint.Color;
+import javafx.scene.text.FontWeight;
 import qupath.lib.common.GeneralTools;
 import qupath.lib.gui.dialogs.Dialogs;
 import qupath.lib.gui.logging.LogManager;
+import qupath.lib.gui.logging.LogManager.LogLevel;
 import qupath.lib.gui.prefs.PathPrefs;
 import qupath.lib.gui.prefs.PathPrefs.FontSize;
 import qupath.lib.gui.prefs.PathPrefs.ImageTypeSetting;
 import qupath.lib.gui.tools.ColorToolsFX;
 import qupath.lib.gui.tools.CommandFinderTools;
 import qupath.lib.gui.tools.CommandFinderTools.CommandBarDisplay;
+import qupath.lib.gui.tools.GuiTools;
 import qupath.lib.gui.prefs.QuPathStyleManager;
 
 /**
@@ -281,12 +290,30 @@ public class PreferencePane {
 				category,
 				"Zoom in when double-clicking on image (if not inside an object) with move tool; zoom out if Alt or Ctrl/Cmd is held down");
 
-		addPropertyPreference(PathPrefs.viewerFontSizeProperty(),
+		addPropertyPreference(PathPrefs.scalebarFontSizeProperty(),
 				FontSize.class,
-				"Viewer font size",
+				"Scalebar font size",
 				category,
-				"Adjust font size for scalebar and location text");
+				"Adjust font size for scalebar");
+		
+		addPropertyPreference(PathPrefs.scalebarFontWeightProperty(),
+				FontWeight.class,
+				"Scalebar font weight",
+				category,
+				"Adjust font weight for the scalebar");
 
+		addPropertyPreference(PathPrefs.scalebarLineWidthProperty(),
+				Double.class,
+				"Scalebar thickness",
+				category,
+				"Adjust line thickness for the scalebar");
+
+		addPropertyPreference(PathPrefs.locationFontSizeProperty(),
+				FontSize.class,
+				"Location text font size",
+				category,
+				"Adjust font size for location text");
+		
 		addPropertyPreference(PathPrefs.useCalibratedLocationStringProperty(), Boolean.class,
 				"Use calibrated location text",
 				category,
@@ -755,12 +782,19 @@ public class PreferencePane {
 		}
 
 	}
+	
+	// We want to reformat the display of these to avoid using all uppercase
+	private static List<Class<?>> reformatTypes = Arrays.asList(
+			FontWeight.class,
+			LogLevel.class
+			);
 
 	/**
 	 * Extends {@link DefaultPropertyEditorFactory} to handle setting directories and creating choice editors.
 	 */
 	public static class PropertyEditorFactory extends DefaultPropertyEditorFactory {
 
+		@SuppressWarnings("unchecked")
 		@Override
 		public PropertyEditor<?> call(Item item) {
 			if (item.getType() == File.class) {
@@ -769,7 +803,21 @@ public class PreferencePane {
 			if (item instanceof ChoicePropertyItem) {
 				return Editors.createChoiceEditor(item, ((ChoicePropertyItem<?>)item).getChoices());
 			}
-			return super.call(item);
+			var editor = super.call(item);
+			if (reformatTypes.contains(item.getType()) && editor.getEditor() instanceof ComboBox) {
+				@SuppressWarnings("rawtypes")
+				var combo = (ComboBox)editor.getEditor();
+				Function<?, String> stringFun = obj -> {
+					var s = Objects.toString(obj);
+					s = s.replaceAll("_", " ");
+					if (Objects.equals(s, s.toUpperCase()))
+						return s.substring(0, 1) + s.substring(1).toLowerCase();
+					return s;
+				};
+				combo.setCellFactory(obj -> GuiTools.createCustomListCell(stringFun));
+				combo.setButtonCell(GuiTools.createCustomListCell(stringFun));
+			}
+			return editor;
 		}
 	}